Cuarta entrega

+++ src/org/joaquinfigueroa/controller/MenuCategoriaProductosController.java
@@ -18,6 +18,7 @@ import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.TextField;
@@ -26,6 +27,7 @@ import org.joaquinfigueroa.dao.Conexion;
 import org.joaquinfigueroa.dto.CategoriaProductoDTO;
 import org.joaquinfigueroa.model.CategoriaProducto;
 import org.joaquinfigueroa.system.Main;
+import org.joaquinfigueroa.utils.SuperKinalAlert;
 
 /**
  * FXML Controller class
@@ -179,9 +181,11 @@ public class MenuCategoriaProductosController implements Initializable {
         }else if(event.getSource() == btnRegresar){
             stage.menuPrincipalView();
         }else if(event.getSource() == btnEliminar){
-            int catProId = ((CategoriaProducto)tblCategoriaProductos.getSelectionModel().getSelectedItem()).getCategoriaProductosId();
-            eliminarCategoriaProducto(catProId);
-            cargarLista();
+            if(SuperKinalAlert.getInstance().mostrarAlertaConfirmacion(405).get() == ButtonType.OK){
+                int catProId = ((CategoriaProducto)tblCategoriaProductos.getSelectionModel().getSelectedItem()).getCategoriaProductosId();
+                eliminarCategoriaProducto(catProId);
+                cargarLista();
+            }
         }else if (event.getSource() == btnBuscar){
             tblCategoriaProductos.getItems().clear();
             if(tfCategoriaProductoId.getText().equals("")){
